+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - New `get_company_details` method in configuration-api v3.6.
+- New `get_license_info` method in agent-api v3.5 web.
 
 ## [0.3.9] - 2024-04-22